Update to new way of vendoring
Instead of symlinking, the new way is to directly move the files out
of the src/ folder up one level.  The previous way of symlinking to
the src/ folder breaks down when I started deploying code out of
my local machine, where the file paths are totally different.
The way the script is used hasn't changed.
README is updated to reflect the change.